requirements.txt
alias_free_torch
cached_conv==2.5.0
einops
huggingface_hub
ipython
librosa
matplotlib
numpy
omegaconf==2.3.0
onnxruntime==1.22.1
packaging==25.0
pandas
pystoi==0.4.1
pytorch_lightning==2.1.0
safetensors==0.6.2
scipy
speechmos
torchcfm==1.0.7
torchdiffeq==0.2.5
torchdyn==1.0.6
torchinfo==1.8.0
tqdm
typing_extensions==4.15.0
wandb
descript-audio-codec
torch-pesq
hydra-core

soundfile
resampy
torchlibrosa
torchvision

laion_clap==1.1.7
transformers<=4.30.2
torchaudio
encodec

optuna

frechet_audio_distance==0.1.2

"huggingface_hub[cli]"